decoding works again

+++ src/genie/quality/calq/qual_decoder.h
@@ -24,7 +24,7 @@ struct DecodingBlock;
 // -----------------------------------------------------------------------------
 
 struct DecodingRead {
-    uint32_t posMin;
+    uint64_t posMin;
     std::string cigar;
 };
